+++ .gitattributes
@@ -1,29 +1,29 @@
 # ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
 
-*.snap                                         	linguist-generated
-/.eslintrc.json                                	linguist-generated
-/.gitattributes                                	linguist-generated
-/.github/pull_request_template.md              	linguist-generated
-/.github/workflows/auto-approve.yml            	linguist-generated
-/.github/workflows/build.yml                   	linguist-generated
-/.github/workflows/release.yml                 	linguist-generated
-/.github/workflows/stale.yml                   	linguist-generated
-/.github/workflows/upgrade-dependencies.yml    	linguist-generated
-/.gitignore                                    	linguist-generated
-/.mergify.yml                                  	linguist-generated
-/.npmignore                                    	linguist-generated
-/.projen/**                                    	linguist-generated
-/.projen/deps.json                             	linguist-generated
-/.projen/jest-snapshot-resolver.js             	linguist-generated
-/.projen/tasks.json                            	linguist-generated
-/lib/__tests__/devapp/cdk.json                 	linguist-generated
-/LICENSE                                       	linguist-generated
-/package.json                                  	linguist-generated
-/src/backend/catalog-builder/catalog-builder.ts	linguist-generated
-/src/backend/discovery/discovery.ts            	linguist-generated
-/src/backend/ingestion/ingestion.ts            	linguist-generated
-/src/backend/transliterator/transliterator.ts  	linguist-generated
-/src/monitoring/certificate-monitor.ts         	linguist-generated
-/src/monitoring/http-get-function.ts           	linguist-generated
-/tsconfig.eslint.json                          	linguist-generated
-/yarn.lock                                     	linguist-generated
+*.snap                                           	linguist-generated
+/.eslintrc.json                                  	linguist-generated
+/.gitattributes                                  	linguist-generated
+/.github/pull_request_template.md                	linguist-generated
+/.github/workflows/auto-approve.yml              	linguist-generated
+/.github/workflows/build.yml                     	linguist-generated
+/.github/workflows/release.yml                   	linguist-generated
+/.github/workflows/stale.yml                     	linguist-generated
+/.github/workflows/upgrade-dependencies.yml      	linguist-generated
+/.gitignore                                      	linguist-generated
+/.mergify.yml                                    	linguist-generated
+/.npmignore                                      	linguist-generated
+/.projen/**                                      	linguist-generated
+/.projen/deps.json                               	linguist-generated
+/.projen/jest-snapshot-resolver.js               	linguist-generated
+/.projen/tasks.json                              	linguist-generated
+/lib/__tests__/devapp/cdk.json                   	linguist-generated
+/LICENSE                                         	linguist-generated
+/package.json                                    	linguist-generated
+/src/backend/catalog-builder/catalog-builder.ts  	linguist-generated
+/src/backend/discovery/discovery.ts              	linguist-generated
+/src/backend/ingestion/ingestion.ts              	linguist-generated
+/src/backend/transliterator/transliterator.ts    	linguist-generated
+/src/monitored-certificate/certificate-monitor.ts	linguist-generated
+/src/monitoring/http-get-function.ts             	linguist-generated
+/tsconfig.eslint.json                            	linguist-generated
+/yarn.lock                                       	linguist-generated
