+++ .gitignore
@@ -56,6 +56,6 @@ yarn-error.log*
 !/src/backend/discovery/discovery.ts
 !/src/backend/ingestion/ingestion.ts
 !/src/backend/transliterator/transliterator.ts
-!/src/monitoring/certificate-monitor.ts
+!/src/monitored-certificate/certificate-monitor.ts
 !/src/monitoring/http-get-function.ts
 !/tsconfig.eslint.json
