+++ src/webapp/index.ts
@@ -7,6 +7,7 @@ import * as s3 from '@aws-cdk/aws-s3';
 import * as s3deploy from '@aws-cdk/aws-s3-deployment';
 import { CfnOutput, Construct } from '@aws-cdk/core';
 import { Domain } from '../api';
+import { MonitoredCertificate } from '../monitored-certificate';
 import { Monitoring } from '../monitoring';
 import { ResponseFunction } from './response-function';
 
@@ -79,7 +80,12 @@ export class WebApp extends Construct {
 
       // Monitor certificate expiration
       if (props.domain.monitorCertificateExpiration ?? true) {
-        props.monitoring.addMonitoredCertificate(props.domain.cert, props.domain.zone.zoneName);
+        const monitored = new MonitoredCertificate(this, 'ExpirationMonitor', {
+          certificate: props.domain.cert,
+          domainName: props.domain.zone.zoneName,
+        });
+        props.monitoring.addHighSeverityAlarm('ACM Certificate Expiry', monitored.alarmAcmCertificateExpiresSoon);
+        props.monitoring.addHighSeverityAlarm('Endpoint Certificate Expiry', monitored.alarmEndpointCertificateExpiresSoon);
       }
     }
